HW01: quiz gamma extend-q3: Avoid pthread for single thread case
In the case of single thread (with option switch -h1),
avoid any pthread initializations and synchronizations.

Unfortunately, this does not really result in performance
improvement which may be due to fact that the overhead of
pthread related calls is relatively inexpensive compared
to other operations such as swapping.

Test results are presented in log.test.skip_pthread_single

Link: https://hackmd.io/@sysprog/linux2023-summer-quiz0